Ensure the need for login for writing operations

+++ src/components/view-movies/view-movies.component.js
@@ -3,6 +3,7 @@
 
 import template from './view-movies.template.html';
 import MoviesService from './../../services/movies/movies.service';
+import UserService from './../../services/user/user.service';
 
 
 class ViewMoviesComponent {
@@ -23,9 +24,10 @@ class ViewMoviesComponent {
 }
 
 class ViewMoviesComponentController{
-    constructor($state,MoviesService){
+    constructor($state,MoviesService,UserService){
         this.$state = $state;
         this.MoviesService = MoviesService;
+        this.UserService = UserService;
 
     }
 
@@ -35,22 +37,31 @@ class ViewMoviesComponentController{
     };
 
     edit (movie) {
-        let _id = movie['_id'];
-        this.$state.go('movieEdit',{ movieId:_id});
+
+        if (this.UserService.isAuthenticated()) {
+            let _id = movie['_id'];
+            this.$state.go('movieEdit',{ movieId:_id});
+        } else {
+            this.$state.go('login',{});
+        }
     };
 
 
     delete(movie) {
-        let _id = movie['_id'];
+        if (this.UserService.isAuthenticated()) {
+            let _id = movie['_id'];
 
-        this.MoviesService.delete(_id).then(response => {
-            this.$state.go('movies',{});
-        });
+            this.MoviesService.delete(_id).then(response => {
+                this.$state.go('movies',{});
+            });
+        } else {
+            this.$state.go('login',{});
+        }
     };
 
 
     static get $inject(){
-        return ['$state', MoviesService.name];
+        return ['$state', MoviesService.name, UserService.name];
     }
 
 }
